Weather logic in separate component
